beamer: chain: make _TransactionFailed and _transact public and move them to util
They are useful on their own and we need this functionality in the
deployment script.
