+++ beamer/util.py
@@ -1,12 +1,38 @@
 import json
 import logging
 import sys
-from typing import List, TextIO
+from typing import Any, List, Optional, TextIO, cast
 
+import requests
 import structlog
 from eth_utils import is_checksum_address, to_checksum_address
+from hexbytes import HexBytes
+from web3.contract import ContractFunction
+from web3.exceptions import ContractLogicError
+from web3.types import TxParams
 
 from beamer.typing import ChainId, ChecksumAddress
+
+
+class TransactionFailed(Exception):
+    def __repr__(self) -> str:
+        return "TransactionFailed(%r)" % self.__cause__
+
+    def __str__(self) -> str:
+        return "transaction failed: %s" % self.cause()
+
+    def cause(self) -> str:
+        return str(self.__cause__)
+
+
+def transact(func: ContractFunction, **kwargs: Any) -> HexBytes:
+    try:
+        txn_hash = func.transact(cast(Optional[TxParams], kwargs))
+    except (ContractLogicError, requests.exceptions.RequestException) as exc:
+        raise TransactionFailed() from exc
+
+    func.web3.eth.wait_for_transaction_receipt(txn_hash)
+    return txn_hash
 
 
 def setup_logging(log_level: str, log_json: bool) -> None:
